% Generated by roxygen2 (4.0.2): do not edit by hand
\name{longFun}
\alias{longFun}
\title{A simple test function for long vectors}
\usage{
longFun(x)
}
\arguments{
\item{x}{A vector of long long integers}
}
\value{
A double vector
}
\description{
Simple test function for long vectors
}
\author{
Dirk Eddelbuettel
}

