% Generated by roxygen2 (4.0.2): do not edit by hand
\name{longlongFun}
\alias{longlongFun}
\title{A simple test function for long long vectors}
\usage{
longlongFun(x)
}
\arguments{
\item{x}{A vector of long long integers}
}
\value{
A double vector
}
\description{
Simple test function for long long vectors
}
\author{
Dirk Eddelbuettel
}

