% Generated by roxygen2 (4.0.2): do not edit by hand
\name{longlongInside}
\alias{longlongInside}
\title{Another simple test function for long long vectors}
\usage{
longlongInside(n)
}
\arguments{
\item{n}{An integer}
}
\value{
A double
}
\description{
Internal use of long long vectors
}
\author{
Dirk Eddelbuettel
}

