\name{RcppLongLong-package}
\alias{RcppLongLong-package}
\alias{RcppLongLong}
\docType{package}
\title{
  Simple test package for long long conversions
}
\description{
  The package does little and serves to tickle some compiler conversions.
}
\details{
  Please see the Rcpp headers.
}
\author{
Dirk Eddelbuettel

Maintainer: Dirk Eddelbuettel <edd@debian.org>
}
% Optionally other standard keywords, one per line,
% from the file KEYWORDS in the R documentation.
\keyword{ package }
\examples{
  # Optional simple examples of the most important functions
}
