% Generated by roxygen2 (4.0.2): do not edit by hand
\name{unsignedlongFun}
\alias{unsignedlongFun}
\title{A simple test function for unsigned long vectors}
\usage{
unsignedlongFun(x)
}
\arguments{
\item{x}{A vector of unsigned long integers}
}
\value{
A double vector
}
\description{
Simple test function for unsigned long vectors
}
\author{
Dirk Eddelbuettel
}

